feat: add UnitDetailsTable component to display inventory unit details in InventoryList
